fix: make directory structure more idiomatic
"pkg" is an idiomatic name for the middle part of a go package's path,
but is an unusual ending. This is already a parsers library and as far
as I know there is no expected need to house many disparate sibling
packages in here, so let's just use the top level.

Move test fixtures into the client package's `testdata` dir, as is also
idiomatic in go.